Revert "qcom-cpufreq: Remove per-cpu workqueue"
This reverts commit ad53079a908da217bede23e4070d4a8f2c7a06d0.

The original commit introduces a bug wherein the HFPLL is reprogrammed
by software with the incorrect assumption that the HFPLL is not the
current source selected by a CPU's mux. Since reprogramming involves
turning off the HFPLL, this results in an unclocked CPU.

This bug is the consequence of a hardware feature wherein a CPU's clock
mux ignores software writes while that CPU is in a certain low power mode.
Therefore if CPU X is changing CPU Y's rate, and CPU Y has entered that
low power mode, CPU X will incorrectly assume that the clock mux has
switched sources.

By reverting the original commit, it is ensured that only CPU Y can
change CPU Y's rate; since it has to be awake in order to execute
the mux-switching software, the error condition is never encountered.

Change-Id: I6adbb8c64dda70f3bf7b339612fd59509ecae217
Signed-off-by: Tianyi Gou <tgou@codeaurora.org>
Signed-off-by: flar2 <asegaert@gmail.com>